Draft: Hiding ChannelOutputStream from the API
