+++ src/java.base/share/classes/java/nio/channels/Channels.java
@@ -41,6 +41,7 @@ import java.nio.channels.spi.AbstractInterruptibleChannel;
 import java.util.Objects;
 import java.util.concurrent.ExecutionException;
 import sun.nio.ch.ChannelInputStream;
+import sun.nio.ch.ChannelOutputStream;
 import sun.nio.cs.StreamDecoder;
 import sun.nio.cs.StreamEncoder;
 
@@ -62,40 +63,6 @@ import sun.nio.cs.StreamEncoder;
 public final class Channels {
 
     private Channels() { throw new Error("no instances"); }
-
-    /**
-     * Write all remaining bytes in buffer to the given channel.
-     * If the channel is selectable then it must be configured blocking.
-     */
-    private static void writeFullyImpl(WritableByteChannel ch, ByteBuffer bb)
-        throws IOException
-    {
-        while (bb.remaining() > 0) {
-            int n = ch.write(bb);
-            if (n <= 0)
-                throw new RuntimeException("no bytes written");
-        }
-    }
-
-    /**
-     * Write all remaining bytes in buffer to the given channel.
-     *
-     * @throws  IllegalBlockingModeException
-     *          If the channel is selectable and configured non-blocking.
-     */
-    private static void writeFully(WritableByteChannel ch, ByteBuffer bb)
-        throws IOException
-    {
-        if (ch instanceof SelectableChannel sc) {
-            synchronized (sc.blockingLock()) {
-                if (!sc.isBlocking())
-                    throw new IllegalBlockingModeException();
-                writeFullyImpl(ch, bb);
-            }
-        } else {
-            writeFullyImpl(ch, bb);
-        }
-    }
 
     // -- Byte streams from channels --
 
@@ -137,65 +104,6 @@ public final class Channels {
     public static OutputStream newOutputStream(WritableByteChannel ch) {
         Objects.requireNonNull(ch, "ch");
         return new ChannelOutputStream(ch);
-    }
-
-    /**
-     * @since 18
-     */
-    public static class ChannelOutputStream extends OutputStream {
-
-        private final WritableByteChannel ch;
-        private ByteBuffer bb;
-        private byte[] bs;       // Invoker's previous array
-        private byte[] b1;
-
-        /**
-         * @param ch The channel wrapped by this stream.
-         */
-        public ChannelOutputStream(WritableByteChannel ch) {
-            this.ch = ch;
-        }
-
-        /**
-         * @return The channel wrapped by this stream.
-         */
-        public WritableByteChannel channel() {
-            return ch;
-        }
-
-        @Override
-        public synchronized void write(int b) throws IOException {
-            if (b1 == null)
-                b1 = new byte[1];
-            b1[0] = (byte) b;
-            this.write(b1);
-        }
-
-        @Override
-        public synchronized void write(byte[] bs, int off, int len)
-                throws IOException
-        {
-            if ((off < 0) || (off > bs.length) || (len < 0) ||
-                ((off + len) > bs.length) || ((off + len) < 0)) {
-                throw new IndexOutOfBoundsException();
-            } else if (len == 0) {
-                return;
-            }
-            ByteBuffer bb = ((this.bs == bs)
-                             ? this.bb
-                             : ByteBuffer.wrap(bs));
-            bb.limit(Math.min(off + len, bb.capacity()));
-            bb.position(off);
-            this.bb = bb;
-            this.bs = bs;
-            Channels.writeFully(ch, bb);
-        }
-
-        @Override
-        public void close() throws IOException {
-            ch.close();
-        }
-
     }
 
     /**
